requirements/base.txt
Django==1.8.2
django-braces==1.4.0
django-model-utils==2.2
pytz==2014.10
django-crispy-forms==1.4.0
django-debug-toolbar==1.3.2
beautifulsoup4==4.3.2
Pillow==2.6.1
psycopg2==2.5.4
django-formtools==1.0